feat: add favicon, fix interface error  and refactor game classes and styles
