+++ src/Pages/Audiocall/audiocall.scss
@@ -44,6 +44,15 @@ button {
   flex-direction: column;
   align-items: center;
 
+  .play-field {
+    display: flex;
+    flex-direction: column;
+    width: 520px;
+    min-height: 410px;
+    position: relative;
+    justify-content: flex-start;
+  }
+
   &__name {
     display: flex;
     justify-content: center;
@@ -151,14 +160,6 @@ button {
 
 .active-level {
   background-color: #58aada;
-}
-
-.play-field {
-  display: flex;
-  flex-direction: column;
-  width: 520px;
-  min-height: 410px;
-  position: relative;
 }
 
 .answers-wrapper {
