+++ src/components/game/interface-game.ts
@@ -1,0 +1,5 @@
+import { Word } from '../../module/apiInterface';
+
+export interface IWordGame extends Word {
+  result?: boolean;
+}
